package: add a postinstall script when lib is distributed via git
Right now, the `lib/defs.js` generated file is only available on
packages published on NPM because it is generated during the `prepare`
phase of the NPM packaging step.

By adding a postinstall step we make sure the file is also generated
by other publishing means (be it local package path or git path for
instance).